Extract slack notification for important bugs to its own util
This will be useful so we can use the function for other tasks.
