+++ projects/github-actions/repo-gardening/src/tasks/triage-issues/index.js
@@ -105,55 +105,18 @@ async function triageIssues( payload, octokit ) {
 			// send a Slack notification.
 			if ( priority === 'TBD' && full_name === 'Automattic/wp-calypso' ) {
 				debug(
-					`triage-issues: #${ number } doesn't have a Priority set. Sending in Slack message to the Kitkat team.`
+					`triage-issues: #${ number } doesn't have a Priority set. Sending in Slack message to the triage team.`
 				);
 				const message = 'New bug missing priority. Please do a priority assessment.';
 				const slackMessageFormat = formatSlackMessage( payload, channel, message );
-				await sendSlackMessage( message, channel, slackToken, payload, slackMessageFormat );
+				await sendSlackMessage( message, channel, payload, slackMessageFormat );
 			}
 		}
 	}
 
-	/*
-	 * Send a Slack Notification if the issue is important.
-	 *
-	 * We define an important issue when meeting all of the following criteria:
-	 * - A bug (includes a "[Type] Bug" label, or a "[Type] Bug" label is added to the issue right now)
-	 * - The issue is still opened
-	 * - The issue is not escalated yet (no "[Status] Priority Review Triggered" label)
-	 * - The issue is either a high priority or a blocker (inferred from the existing labels or from the issue body)
-	 * - The issue is not already set to another priority label (no "[Pri] High", "[Pri] BLOCKER", or "[Pri] TBD" label)
-	 */
-
-	const isEscalated = await hasEscalatedLabel( octokit, ownerLogin, name, number, action, label );
-
-	const highPriorityIssue = priority === 'High' || priorityLabels.includes( '[Pri] High' );
-	const blockerIssue = priority === 'BLOCKER' || priorityLabels.includes( '[Pri] BLOCKER' );
-
-	const hasOtherPriorityLabels = priorityLabels.some( priLabel =>
-		/^\[Pri\] (?!High|BLOCKER|TBD)/.test( priLabel )
-	);
-
-	if (
-		isBugIssue &&
-		state === 'open' &&
-		! isEscalated &&
-		( highPriorityIssue || blockerIssue ) &&
-		! hasOtherPriorityLabels
-	) {
-		const message = `New ${
-			highPriorityIssue ? 'High-priority' : 'Blocker'
-		} bug! Please check the priority.`;
-		const slackMessageFormat = formatSlackMessage( payload, channel, message );
-		await sendSlackMessage( message, channel, slackToken, payload, slackMessageFormat );
-
-		debug( `triage-issues: Adding a label to issue #${ number } to show that Kitkat was warned.` );
-		await octokit.rest.issues.addLabels( {
-			owner: ownerLogin,
-			repo: name,
-			issue_number: number,
-			labels: [ '[Status] Priority Review Triggered' ],
-		} );
+	// Send a Slack notification if the issue is important.
+	if ( isBugIssue ) {
+		await notifyImportantIssues( octokit, payload );
 	}
 }
 module.exports = triageIssues;
